refactoring of the code for better modularity
