+++ intunlu/train.py
@@ -6,13 +6,13 @@ import random
 import time
 import torch
 
-from datasets import load_dataset
 from pytorch_lightning.callbacks.early_stopping import EarlyStopping
 from pytorch_lightning.loggers import TensorBoardLogger
-from rouge_score import rouge_scorer
 from transformers import T5Tokenizer, BartTokenizer
 
 from intunlu.finetunning import SummaryDataModule, SummarizerModel
+from intunlu.evaluate import evaluate
+from intunlu.utils import setup_logger, load_data
 
 
 def train(
@@ -92,88 +92,6 @@ def train(
     logging.info(f'(Took {time.time() - s} seconds.)')
     with open(f'results_{random_state}.pkl', 'wb')as f:
         pickle.dump(results, f)
-
-
-def load_data(max_num_samples=None):
-    # load datasets and trim them down as needed
-    datasets = {}
-    for ds in ['train', 'validation', 'test']:
-        dataset = load_dataset('xsum', split=ds)
-        N = len(dataset)
-        logging.info(f'{ds} dataset: initial # of samples is {N}.')
-        if max_num_samples is not None:
-            n = min(max_num_samples, N)
-        else:
-            n = N
-        logging.info(f'{ds} dataset: will effectively use {n} samples.')
-        dataset = list(zip(dataset['document'], dataset['summary']))
-        dataset = random.sample(dataset, n)
-        datasets[ds] = {
-            'document': ['summarize: ' + e[0].replace('\n', ' ') for e in dataset],
-            'summary': [e[1] for e in dataset]
-        }
-
-    return datasets
-
-
-def evaluate(model, dataset):
-
-    scorer = rouge_scorer.RougeScorer(['rouge1', 'rouge2', 'rougeL'], use_stemmer=False)
-
-    results = {
-        'reference_summary': [],
-        'predicted_summary': [],
-        'rouge1': [],
-        'rouge2': [],
-        'rougeL': []
-    }
-
-    device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
-    model.model.to(device)
-    for i in range(len(dataset['document'])):
-        document = model.tokenizer(
-            'summarize: ' + dataset['document'][i],
-            padding='longest',
-            truncation=True,
-            return_tensors='pt'
-        )
-        with torch.no_grad():
-            pred = model.model.generate(
-                input_ids=document['input_ids'].to(device),
-                attention_mask=document['attention_mask'].to(device),
-                use_cache=True,
-                decoder_start_token_id=model.tokenizer.pad_token_id,
-                num_beams=1,  # greedy search
-                early_stopping=True
-            )
-            pred = model.tokenizer.convert_ids_to_tokens(pred[0], skip_special_tokens=True)
-            pred = model.tokenizer.convert_tokens_to_string(pred).replace(' . ', '. ')
-            if i < 2:
-                logging.info('The input document:')
-                logging.info(dataset['document'][i])
-                logging.info('The reference summary:')
-                logging.info(dataset['summary'][i])
-                logging.info('The predicted summary:')
-                logging.info(pred)
-        results['reference_summary'].append(dataset['summary'][i])
-        results['predicted_summary'].append(pred)
-        scores = scorer.score(results['reference_summary'][0], results['predicted_summary'][0])
-        results['rouge1'].append(scores['rouge1'][1])
-        results['rouge2'].append(scores['rouge2'][1])
-        results['rougeL'].append(scores['rougeL'][2])
-
-    return results
-
-
-def setup_logger(random_state):
-    logging.basicConfig(
-        level=logging.INFO,
-        format="%(asctime)s [%(threadName)-12.12s] [%(levelname)-5.5s]  [%(filename)s:%(lineno)s] [%(funcName)20s()] %(message)s",
-        handlers=[
-            logging.FileHandler(f"{random_state}.log"),
-            logging.StreamHandler()
-        ]
-    )
 
 
 if __name__ == '__main__':
